requirements.txt
attrs==23.2.0
black==24.4.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
